requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Audio processing
pydub>=0.25.1
elevenlabs>=0.2.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Transcription and subtitles
whisper-timestamped>=1.14.0
openai-whisper>=20230918

# Additional utilities
numpy>=1.24.0
Pillow>=10.0.0